Add partial match repo lookup / search
This improves user-friendliness by allowing the user to type just the name of the repository
or even just a part of the name to find the repository.
The username of the repo's owner is also supported for lookup
as well as the server name / netloc for even easier blind searches.
Typing "github" will now find all repos from GitHub,
unless a repository called "github" or
a repository from a user "github" is successfully analyzed.